requirements.txt
faster_whisper==1.2.1
Flask==3.1.3
numpy==2.4.2
openai==2.21.0
python-dotenv==1.2.1
sounddevice==0.5.5
